refactor(applications): move subscription types and constants to packages/applications
