+++ apps/api/src/server/applications/subscriptions/subscriptions.validators.js
@@ -1,6 +1,7 @@
 import { composeMiddleware } from '@pins/express';
 import { body, param, query } from 'express-validator';
 import { validationErrorHandler } from '#middleware/error-handler.js';
+import { Subscription } from '@pins/applications/lib/application/subscription.js';
 
 export const validateGetSubscription = composeMiddleware(
 	query('caseReference').notEmpty().withMessage(`caseReference is required`),
@@ -78,16 +79,10 @@ function validateStartBeforeEnd(startDate, endDate) {
 	return start < end;
 }
 
-const subscriptionTypes = [
-	'allUpdates',
-	'applicationSubmitted',
-	'applicationDecided',
-	'registrationOpen'
-];
 /**
  * @param {string} v
  * @returns {boolean}
  */
 function validateSubscriptionType(v) {
-	return subscriptionTypes.includes(v);
+	return Subscription.TypeList.includes(v);
 }
