+++ packages/applications/types/index.d.ts
@@ -4,3 +4,4 @@ export * from './file-folders';
 export * from './project-update';
 export * from './representation';
 export * from './s51-advice';
+export * from './subscription';
